requirements.txt
langchain>=0.0.200
langchain-anthropic>=0.1.0
pandas>=1.4.0
python-dotenv>=0.21.0 